Fixed update issue, linter
